+++ zerver/tests/test_realm_domains.py
@@ -2,9 +2,10 @@ import orjson
 from django.core.exceptions import ValidationError
 from django.db.utils import IntegrityError
 
+from zerver.actions.realm_domains import do_change_realm_domain, do_remove_realm_domain
 from zerver.actions.realm_settings import do_set_realm_property
 from zerver.actions.users import do_change_user_role
-from zerver.lib.actions import do_change_realm_domain, do_create_realm, do_remove_realm_domain
+from zerver.lib.actions import do_create_realm
 from zerver.lib.domains import validate_domain
 from zerver.lib.email_validation import email_allowed_for_realm
 from zerver.lib.test_classes import ZulipTestCase
